# src/routes/swagger.json
{
	"components": {
...
		"schemas": {
			"IResponse": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double",
						"description": "This is the code of response you consult the code list for known what this code meaning"
					},
					"data": {
						"description": "This is the data (Optionnal)"
					},
					"message": {
						"type": "string",
						"description": "This is the message of the request (Optionnal)"
					},
					"token": {
						"type": "string",
						"description": "This is the token givin on login (Optionnal)"
					}
				},
				"required": [
					"code",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateCommentParams": {
				"properties": {
					"name": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"email": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"postId": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"message": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					}
				},
				"required": [
					"name",
					"email",
					"postId",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCommentParams": {
				"properties": {
					"id": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"message": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePostParams": {
				"properties": {
					"title": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"slug": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"categoryId": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"cover": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"content": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"state": {
						"type": "string",
						"enum": [
							"DRAFT",
							"PUBLISHED",
							"ARCHIVED"
						],
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					}
				},
				"required": [
					"title",
					"slug",
					"categoryId",
					"cover",
					"content",
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdatePostParams": {
				"properties": {
					"title": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"slug": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"categoryId": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"cover": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"content": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"state": {
						"type": "string",
						"enum": [
							"DRAFT",
							"PUBLISHED",
							"ARCHIVED"
						],
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					},
					"id": {
						"type": "string",
						"description": "This is a description of a special string property",
						"example": "password",
						"minLength": 3,
						"maxLength": 20
					}
				},
				"required": [
					"title",
					"slug",
					"categoryId",
					"cover",
					"content",
					"state"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "token",
				"in": "header",
				"scopes": {
					"root": "can do",
					"admin": "can do",
					"user": "can do"
				}
			}
		}
	},
	"info": {
		"title": "darwin",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/comment/create": {
			"post": {
				"operationId": "CreateComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Comment"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCommentParams"
							}
						}
					}
				}
			}
		},
		"/comment/update": {
			"post": {
				"operationId": "UpdateComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Comment"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCommentParams"
							}
						}
					}
				}
			}
		},
		"/comment/delete": {
			"post": {
				"operationId": "DeleteComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Comment"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "string"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/comment/getAll": {
			"get": {
				"operationId": "GetAllComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Comment"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "postId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/post/create": {
			"post": {
				"operationId": "CreatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePostParams"
							}
						}
					}
				}
			}
		},
		"/post/update": {
			"post": {
				"operationId": "UpdatePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePostParams"
							}
						}
					}
				}
			}
		},
		"/post/delete": {
			"post": {
				"operationId": "DeletePost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"id": {
										"type": "string"
									}
								},
								"required": [
									"id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/post/getAll": {
			"get": {
				"operationId": "GetAllPost",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Post"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}